requirements.txt
dnspython==2.8.0
numpy==2.2.6
pandas==2.3.3
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.3
